backend/requirements.txt
# DocuMind AI Backend Dependencies
# Phase 1: Ingestion Pipeline

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
pymongo==4.6.1
dnspython==2.4.2

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for future phases)
httpx==0.26.0

# =============================================================================
# Phase 2: Embeddings & Retrieval
# =============================================================================

# Vector Operations
numpy==1.26.3

# OpenAI Embeddings
openai==1.12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
